+++ package.json
@@ -10,6 +10,7 @@
   "description": "",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.4.7",
     "express": "^4.21.1",
     "pg": "^8.13.1",
     "router": "^1.3.8"
